Add function in intermediate compiler to prevent double-counting of margin
When the AST has an if-within branch only the largest margin in the
branch should be counted, otherwise an unreasonably high amount would
in some contracts be used as margin.

This commit adds a naive way of achieving this (with exponential running
time in the height of the AST). It also adds a unit test checking that
the implementation is sane.