Add the calculated activateMap data structure to the intermediate compiler type
This commit adds the activateMap data structure to the IntermediateContract
such that it is ready to be used in the EVM compiler. The EVM compiler will
need this data map when constructing the Activate function. The data in the
activateMap will be used as arguments in the list of transferFrom calls.