Allow getFunctionCallEvm to take DC's own address as function argument
This commit adds a functionality such that the function call
can pass the compiled contract's address to the callee in a
function call. Previously the getFunctionCallEvm only accepted
Word256 in the function call argument list; now it can handle
both Word256 and the contract's own address.

The new functionality has not been tested, only a regression
test has been performed through get_tools ensuring that the
existing functionality of getFunctionCallEvm has not been destroyed.
This regression test is called "run_example_dagger" in geth_tools
and it is a test that checks that the correct function call to
an observable contract is created by the getFunctionCallEvm.