Add non-completed getActivate function
The getActivate function returns the code that executes the
transferFrom calls which are needed to lock the margin funds
on the DC address. This function is not yet completed since
a check of the returned boolean indicating whether the
transferFrom call was successful or not has not yet been added.