Reduce gas usage when deploying DCs by not storing TC-specific values
This commit removes from the contructor the writing to storage of
TC-specific values such as the _delay, _toAddress etc. which was
previously done by the constructor. Instead these values are now
hardcoded in the produced code upon compilation.

This saves on the gas used when deploying contracts since writing to
storage is expensive.