Changed test framework to Hspec, improved pretty-printer
The test framework is now Hspec, so HUnit and QuickCheck tests are
displayed more nicely together. (Another candidate framework that has
this feature is Testy, and I'm not sure which one is better, but both
are popular.)

The pretty-printer was improved so it doesn't print unnecessary parens.
Unfortunately I didn't know the associativity rules, so right now it
assumes that all binary operators are associative. For this reason, it
fails on stuff like "a < b < c" which the parser seems to disallow. So
it needs a bit more tweaking before the test is enabled.