Add simple function call functionality to generated code
This commit adds a simple calling convention to the produced EVM
allowing code to be reused to a larger degree than it is today.

The stack is used to store the return address and it is the callee's
responsibility to clear its own stack before calling RETURN which
is simply just JUMP (which sets PC to the value stored at the top
of the stack).