Move safeMul to subroutine, fix safeMul duplicate label bug
In 'evmCompile', multiple calls to runCompiler do not share the unique
label generator, so we risk duplicate labels.

We think safeMul was used in multiple parts of 'evmCompile' that reset
the unique name counter in each iteration. Making a safeMul subroutine
giving it a constant, global label removes the problem in safeMul.

Co-authored-by: Thorkil Værge <thorkilk@gmail.com>