Change setBalance subroutine to accept account argument
This extra arugmnet is needed to implement a transfer function
(safeTransferFrom) as the balane in these calls are updated for two
parties: the _from *and*  the _to address. Previously, the setBalance
subroutine could only update caller's balance.

All uses have been fixed to now include this extra argument, and the
functionality has been verified in the test suite in geth_tools.