Get byte values correctly in safeBatchTransferFrom function
This commit reads the pointer to the two arrays correctly from
CALLDATALOAD instead of relying on magic variables. The ABI encoding of
arrays put a pointer to the array where this pointer is a byte address
of the beginning of the array (excluding the function signature) where
the array is placed in the function signature. In other words: In place
of the array argument is a pointer. Add four to this pointer and you
will get the byte address in I_d (CALLDATALOAD/ROM) where the length of
the array is specified. Following the length of the array are the
elements.

Co-authored-by: Simon Shine <shreddedglory@gmail.com>