Change order in jump table to minimize gas costs
The rule is that the non-constant methods are placed order from what we
believe is the most commonly used method to the least used. To reduce
the total amount of code, three of the derivative's methods, `activate`,
`mint`, and `burn` are placed last since they can then share the code
for an input validation step that ensures that the input argument (for
all three methods) is strictly postive.