Move CALLER check in transfer methods from subroutine to its caller
This moves the check that the caller is approved to withdraw from the
sender (_from) address from the subroutine to the two methods where this
subroutine is invoked. This is done to reduce the gas cost of
`safeBatchTransferFrom` as this check will now only be performed once,
and not N times where N is the length of the array arguments.

This should make the contract slightly more expensive to deploy but
cheaper to run in some cases and we optimize for run time, not for
deploy.

The gas cost saved here i significant since the CALLER check requires a
read from storage that costs 5000 gas.