Add check that parties are numberede sequentially #21
This is done in preparation for ensuring that funds are not stuck in the
DC. Our solution to prevent funds being stuck in DC requires all party
payouts from 1 to N to be evaluated before the payout to tokenID 0 is
made. tokenID 0 will be a party that is added (if needed) to represent
any remainig position iff the contract is not guaranteed to pay out all
funds without a tokenID 0.