Skip to end if evaluated payout is zero
Issue: #32

If the evaluated payout for a TransferCall is zero, set PT balance to
zero and end the TransferCall evaluation. This is a significant
gas-saving optimization and it is the 1st in two commits where the goal
is to prevent calls to external settlement asset ERC20 contracts with a
value of 0.

Notice that the position of this check ensures that it catches both the
case where the TC value has already been evaluated **and** where it has
**just** been evaluated in this VM invocation (blockchain transcation).

Tests with and without this optimization were run for two of the tests
in our integration test suite `geth_tools`. We consider the gas cost for
repeated calls to `pay`

PayChecks test:
without optimization:
52483, 40109, 36720, 48418
with this optimization:
52483, 35042, 31653, 38284

EnsurePt0PayoutNotBlocked3 test:
without optimization:
97631, 42947
with optimization:
87497, 26330

This optimization thus saves significatantly more than 10 % of the
consumed gas for `pay`. The other endpoints are unaffected by it.