Use constant folding optimization
The constant folding was written but it was not invoked from the Main
file, so its optimization was not applied to the generated contracts.
This saves 60/244 gas on the first invocation of the
PayChecks/EnsurePt0PayoutNotBlocked3 tests. On later invocations it does
not save anything as the value of the scale expression is always only
invoked once. This should also save some on the deployment gas cost as
we save the boilerplate of a function call/subroutine invocation of
SafeMul but since the SafeMul subroutine is always included in the
deployed contracts, this amount is rather small, probably less than 0x0a
bytes.

The scale expression previously always looked like:
`MultExp (Lit (IntVal 1)) expr `
since this is how the intermediate compiler constructs the
expression recursively. Applying constant folding removes this
superfluous multiplication with 1, so it saves a call to SafeMul.