Only run CI on pushes to master
It appears that enabling CI for all branches and all PRs causes double
runs on PRs. An example of this can be seen e.g. in #36. This commit
changes so that only direct pushes to master, as well as any PRs
(including ones that merge with other targets than master) trigger CI.