Remove superfluous transferCall ID
This TCID value is not needed as all transfer calls are already uniquely
identified by their party token (which currently is held in the _to
parameter of the intermediate contract). This code removes this ID from
the code base. The only non-trivial change is to the expression
`subtractEvaluatedPTValuesFromSaAmounts` where a calculation of the
payback streams to the PT0 holder is calculated. This loops over all
party token IDs from the max value to 1 and calculates the how much
should be paid back to PT0. Previously the condition to check if the
loop was done was tcid <= 0; now the condition is ptid == 0, this saves
three instructions. This function also necessitated that a subroutine
was changed, it previously mapped from map from TCID to SAID but now
maps from PTID to SAID.